Fixed bug where users could be viewed without auth
If req.user exists, passport assumes that there's an active session.
When you query /users/:id, there was an app.param() call that preloads
the user with that ID into req.user. It's now loaded into req.userID,
and I believe I've fixed all usages of that in the user controller.