Lazily load exports to remove cyclic deps
Turns out if you have a two-way relationship in a DB, and want to reuse
your models, you can't just have each require the other AND get
something out of that import's "module.exports", like this:

    var Playlist = require("./playlist").model

This loads the dependency in full, and all its requirements. So, if
there's a cyclic dependency, it'll cause errors in your code. Instead,
this sort of pattern works nicely:

    // At the top
    var Playlist = require("./playlist")

    // in the body of your code somewhere, for example
    this.belongsTo(Playlist.model, "PlaylistID");